Added Xcode 5 test action support. This allows to have a dedicated build track for running unit/integration tests. - Made XCodeBuild abstract, and subclassed by XcodeBuilder and XcodeTest - Extracted XcodeDescriptor to an abstract class, for reuse by two newly created classes - Added relevant Jelly config for XcodeTester
