+++ src/main/resources/au/com/rayh/Messages.properties
@@ -37,7 +37,8 @@ XCodeBuilder.fetchingCFBundleVersion=Fetching technical version number (CFBundle
 XCodeBuilder.invokeXcodebuild=Going to invoke xcodebuild:
 XCodeBuilder.packagingIPA=Packaging IPA
 XCodeBuilder.workingDir=Working directory is {0}.
-XCodeBuilder.xcode=Xcode
+XCodeBuilder.xcode=Xcode Build
+XCodeBuilder.xcodeTest=Xcode Test
 XCodeBuilder.xcodeVersionNotFound=Check your Xcode installation. Jenkins cannot retrieve its version.
 XCodeBuilder.CFBundleShortVersionStringNotFound=No marketing version found (CFBundleShortVersionString).
 XCodeBuilder.CFBundleShortVersionStringFound=Found marketing version (CFBundleShortVersionString): {0}.
