+++ src/main/resources/au/com/rayh/XCodeTester/config.jelly
@@ -1,0 +1,142 @@
+<!--
+  ~ The MIT License
+  ~
+  ~ Copyright (c) 2011 Ray Yamamoto Hilton
+  ~
+  ~ Permission is hereby granted, free of charge, to any person obtaining a copy
+  ~ of this software and associated documentation files (the "Software"), to deal
+  ~ in the Software without restriction, including without limitation the rights
+  ~ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+  ~ copies of the Software, and to permit persons to whom the Software is
+  ~ furnished to do so, subject to the following conditions:
+  ~
+  ~ The above copyright notice and this permission notice shall be included in
+  ~ all copies or substantial portions of the Software.
+  ~
+  ~ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+  ~ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+  ~ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+  ~ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+  ~ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+  ~ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+  ~ THE SOFTWARE.
+  -->
+<j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:html="jelly:html">
+	<![CDATA[
+		<script type="text/javascript">
+		<!--
+			var showOrHideKeychainAttributeInput = function(uuid) {
+				var defaultChosen = (document.getElementById("global-keychain-select-" + uuid).selectedIndex == 0);
+				
+				document.getElementById("" + uuid).style.display = defaultChosen ? "block" : "none";
+			};
+		-->
+		</script>
+	]]>
+
+	<!-- General build settings -->
+	<f:section title="${%General build settings}">
+	    
+	    <f:entry title="${%Target}" field="target"
+	      description="Leave empty for all targets">
+	        <f:textbox/>
+	    </f:entry>
+	    
+		<f:advanced title="Settings">
+		    <f:entry title="${%Clean before build?}" field="cleanBeforeBuild"
+		      description="This will delete the build directories before invoking the build.">
+		        <f:checkbox title="Yes" />
+		    </f:entry>
+		   
+		    <f:entry title="${%Allow failing build results?}" field="allowFailingBuildResults"
+		      description="Checking this option will prevent this build step from failing if xcodebuild exits with a non-zero return code.">
+		        <f:checkbox title="Yes" />
+		    </f:entry>
+		    
+		    <f:entry title="${%Configuration}" field="configuration"
+		      description="This is the name of the configuration as defined in the Xcode project.">
+		        <f:textbox default="Release"/>
+		    </f:entry>
+
+		    <f:entry name="testDevice" title="${%Device}" field="testDevice">
+                 <f:select />
+            </f:entry>
+
+             <f:entry name="testOsVersion" title="${%iOS Version}" field="testOsVersion">
+                <f:select />
+            </f:entry>
+
+            <f:entry title="${%Reset simulator?}" field="resetSimulator"
+                    description="This will reset the simulator before running the tests.">
+                <f:checkbox title="Yes" />
+            </f:entry>
+
+
+		</f:advanced>
+	</f:section>
+	
+	<!-- Advanced build options -->
+	<f:section title="${%Advanced Xcode build options}">
+		<f:advanced title="Advanced build settings">
+		
+		    <f:entry title="${%Xcode Schema File}" field="xcodeSchema"
+		      description="Only needed if you want to compile for a specific schema instead of a target.">
+		        <f:textbox />
+		    </f:entry>
+		
+		    <f:entry title="${%SDK}" field="sdk"
+		      description="Leave empty for default SDK">
+		        <f:textbox/>
+		    </f:entry>
+		
+		    <f:entry title="${%SYMROOT}" field="symRoot"
+		      description="Leave empty for default SYMROOT">
+		        <f:textbox/>
+		    </f:entry>
+		
+		    <f:entry title="${%Custom xcodebuild arguments}" field="xcodebuildArguments"
+		      help="/plugin/xcode/help-xcodebuildArguments.html">
+		        <f:textbox name="xcode.xcodebuildArguments" value="${instance.xcodebuildArguments}" default=""/>
+		    </f:entry>
+		
+		    <f:entry title="${%Xcode Workspace File}" field="xcodeWorkspaceFile"
+		      description="Only needed if you want to compile a workspace instead of a project.">
+		        <f:textbox />
+		    </f:entry>
+		
+		    <f:entry title="${%Xcode Project Directory}" field="xcodeProjectPath"
+		      description="Relative path within the workspace that contains the xcode project file(s).">
+		        <f:textbox />
+		    </f:entry>
+		
+		    <f:entry title="${%Xcode Project File}" field="xcodeProjectFile"
+		      description="Only needed if there is more than one project file in the Xcode Project Directory">
+		        <f:textbox />
+		    </f:entry>
+		
+		    <f:entry title="${%Build output directory}" field="configurationBuildDir"
+		         description="The value to use for CONFIGURATION_BUILD_DIR setting.">
+		           <f:textbox/>
+		    </f:entry>
+		    
+		</f:advanced>
+
+	</f:section>
+	
+	<f:section title="${%Versioning}">
+	    <f:entry>
+	        <f:optionalBlock inline="true" checked="${instance.provideApplicationVersion == true}" name="provideApplicationVersion" title="${%Provide version number and run avgtool?}" description="Checking this option will run avgtool and update the CFBundleVersion and CFBundleVersionShortString.">
+	      <f:entry title="${%Marketing version}" field="cfBundleShortVersionStringValue"
+	            description="The value to use for CFBundleShortVersionString. Leave blank to use project's marketing number.">
+	            <f:textbox />
+	          </f:entry>
+	
+	      <f:entry title="${%Technical version}" field="cfBundleVersionValue"
+	        description="The value to use for CFBundleVersion. Leave blank to use project's technical number.">
+	          <f:textbox />
+	      </f:entry>
+	        </f:optionalBlock>
+	    </f:entry>
+	</f:section>
+
+</j:jelly>
